Added some interesting graphical features and a reset all button.
Currently, the implementation of reset all is pretty inneficient. It
programatically clicks each of the other buttons, and each button, when
clicked, replots. Ideally, i need to delay replotting when the reset all
button is clicked. So use it at your own risk.

Also, these changes were hacked in in a fury of activity. I'll need to
do some cleanup later.